Project/LoL/pyproject.toml
[tool.poetry]
name = "lol-data-analysis"
version = "0.1.0"
description = ""
authors = ["uowol <alllhvhla@gmail.com>"]
readme = "README.md"


[tool.black]
line-length = 88
target-version = ['py312']
skip-string-normalization = false


[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true


[tool.poetry.dependencies]
python = ">=3.10,<3.13"
dotenv = "^0.9.9"
requests = "^2.32.3"
aiohttp = "^3.11.13"
aiofiles = "^24.1.0"
duckdb = "^1.2.1"
numpy = "^2.2.4"
pandas = "^2.2.3"
psycopg2 = "^2.9.10"
pyaml = "^25.1.0"
pydantic = "^2.11.2"
